pyproject.toml
[project]
name = "legleux-generate-ledger"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Michael Legleux", email = "legleux@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "base58>=2.1.1",
    "click>=8.2.1",
    "httpx>=0.28.1",
    "platformdirs>=4.3.8",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "ruamel-yaml>=0.18.14",
    "typer>=0.17.4",
    "xrpl-py>=4.2.0",
]

[project.scripts]

gen = "generate_ledger.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/generate_ledger"]

[tool.hatch.build.targets.sdist]
include = [
  "src/generate_ledger",
  "pyproject.toml",
  "README.md",
#   "LICENSE*",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E","F","I","W","B","UP","ISC","PL","RUF"]
ignore = ["E203"]

[dependency-groups]
dev = [
  {include-group = "lint"},
  {include-group = "test"}
]
lint = [
    "ruff>=0.12.11",
]
test = [
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
# minversion = "8.0"
# addopts = "-q"
addopts = "-rP"
# addopts = "-rx"
# addopts = "-qrP"
# testpaths = ["tests"]
